+++ test/docs/schematypes.test.js
@@ -10,8 +10,8 @@ describe('schemaTypes', function() {
     db = mongoose.createConnection('mongodb://localhost:27017/mongoose_test');
   });
 
-  after(function(done) {
-    db.close(done);
+  after(async function() {
+    await db.close();
   });
 
   /**