+++ test/model.findOneAndRemove.test.js
@@ -22,8 +22,8 @@ describe('model: findOneAndRemove:', async function() {
     db = start();
   });
 
-  after(function(done) {
-    db.close(done);
+  after(async function() {
+    await db.close();
   });
 
   beforeEach(() => db.deleteModel(/.*/));