+++ test/query.cursor.test.js
@@ -19,8 +19,8 @@ describe('QueryCursor', function() {
     db = start();
   });
 
-  after(function(done) {
-    db.close(done);
+  after(async function() {
+    await db.close();
   });
 
   beforeEach(() => db.deleteModel(/.*/));