+++ test/schema.alias.test.js
@@ -18,8 +18,8 @@ describe('schema alias option', function() {
     db = start();
   });
 
-  after(function(done) {
-    db.close(done);
+  after(async function() {
+    await db.close();
   });
 
   beforeEach(() => db.deleteModel(/.*/));