+++ test/schema.select.test.js
@@ -17,8 +17,8 @@ describe('schema select option', function() {
     db = start();
   });
 
-  after(function(done) {
-    db.close(done);
+  after(async function() {
+    await db.close();
   });
 
   beforeEach(() => db.deleteModel(/.*/));